+++ kibitzr/__init__.py
@@ -2,4 +2,4 @@
 
 __author__ = 'Peter Demin'
 __email__ = 'kibitzrrr@gmail.com'
-__version__ = '3.1.8'
+__version__ = '4.0.0b2'